murmur3: switch to unsigned types
C++ doesn't define overflow on signed types, so use unsigned types instead.
Luckily all right shifts were unsigned anyway.

Some signed extension was happening (handling remainders after processing
8-byte chunks) but should still be there.

Caught by debug build.

+++ utils/murmur_hash.cc
@@ -26,19 +26,19 @@ namespace utils {
 
 namespace murmur_hash {
 
-int32_t hash32(const bytes &data, int32_t offset, int32_t length, int32_t seed)
+uint32_t hash32(const bytes &data, uint32_t offset, uint32_t length, uint32_t seed)
 {
-    int32_t m = 0x5bd1e995;
-    int32_t r = 24;
+    uint32_t m = 0x5bd1e995;
+    uint32_t r = 24;
 
-    int32_t h = seed ^ length;
+    uint32_t h = seed ^ length;
 
-    int32_t len_4 = length >> 2;
+    uint32_t len_4 = length >> 2;
 
-    for (int32_t i = 0; i < len_4; i++)
+    for (uint32_t i = 0; i < len_4; i++)
     {
-        int32_t i_4 = i << 2;
-        int32_t k = data[offset + i_4 + 3];
+        uint32_t i_4 = i << 2;
+        uint32_t k = data[offset + i_4 + 3];
         k = k << 8;
         k = k | (data[offset + i_4 + 2] & 0xff);
         k = k << 8;
@@ -53,22 +53,22 @@ int32_t hash32(const bytes &data, int32_t offset, int32_t length, int32_t seed)
     }
 
     // avoid calculating modulo
-    int32_t len_m = len_4 << 2;
-    int32_t left = length - len_m;
+    uint32_t len_m = len_4 << 2;
+    uint32_t left = length - len_m;
 
     if (left != 0)
     {
         if (left >= 3)
         {
-            h ^= (int32_t) data[offset + length - 3] << 16;
+            h ^= (uint32_t) data[offset + length - 3] << 16;
         }
         if (left >= 2)
         {
-            h ^= (int32_t) data[offset + length - 2] << 8;
+            h ^= (uint32_t) data[offset + length - 2] << 8;
         }
         if (left >= 1)
         {
-            h ^= (int32_t) data[offset + length - 1];
+            h ^= (uint32_t) data[offset + length - 1];
         }
 
         h *= m;
@@ -81,52 +81,52 @@ int32_t hash32(const bytes &data, int32_t offset, int32_t length, int32_t seed)
     return h;
 }
 
-int64_t hash2_64(const bytes &key, int32_t offset, int32_t length, int64_t seed)
+uint64_t hash2_64(const bytes &key, uint32_t offset, uint32_t length, uint64_t seed)
 {
-    int64_t m64 = 0xc6a4a7935bd1e995L;
-    int32_t r64 = 47;
+    uint64_t m64 = 0xc6a4a7935bd1e995L;
+    uint32_t r64 = 47;
 
-    int64_t h64 = (seed & 0xffffffffL) ^ (m64 * length);
+    uint64_t h64 = (seed & 0xffffffffL) ^ (m64 * length);
 
-    int32_t lenLongs = length >> 3;
+    uint32_t lenLongs = length >> 3;
 
-    for (int32_t i = 0; i < lenLongs; ++i)
+    for (uint32_t i = 0; i < lenLongs; ++i)
     {
-        int32_t i_8 = i << 3;
+        uint32_t i_8 = i << 3;
 
-        int64_t k64 =  ((int64_t)  key[offset+i_8+0] & 0xff)      + (((int64_t) key[offset+i_8+1] & 0xff)<<8)  +
-                (((int64_t) key[offset+i_8+2] & 0xff)<<16) + (((int64_t) key[offset+i_8+3] & 0xff)<<24) +
-                (((int64_t) key[offset+i_8+4] & 0xff)<<32) + (((int64_t) key[offset+i_8+5] & 0xff)<<40) +
-                (((int64_t) key[offset+i_8+6] & 0xff)<<48) + (((int64_t) key[offset+i_8+7] & 0xff)<<56);
+        uint64_t k64 =  ((uint64_t)  key[offset+i_8+0] & 0xff)      + (((uint64_t) key[offset+i_8+1] & 0xff)<<8)  +
+                (((uint64_t) key[offset+i_8+2] & 0xff)<<16) + (((uint64_t) key[offset+i_8+3] & 0xff)<<24) +
+                (((uint64_t) key[offset+i_8+4] & 0xff)<<32) + (((uint64_t) key[offset+i_8+5] & 0xff)<<40) +
+                (((uint64_t) key[offset+i_8+6] & 0xff)<<48) + (((uint64_t) key[offset+i_8+7] & 0xff)<<56);
 
         k64 *= m64;
-        k64 ^= (uint64_t)k64 >> r64;
+        k64 ^= k64 >> r64;
         k64 *= m64;
 
         h64 ^= k64;
         h64 *= m64;
     }
 
-    int32_t rem = length & 0x7;
+    uint32_t rem = length & 0x7;
 
     switch (rem)
     {
     case 0:
         break;
     case 7:
-        h64 ^= (int64_t) key[offset + length - rem + 6] << 48;
+        h64 ^= (uint64_t) key[offset + length - rem + 6] << 48;
     case 6:
-        h64 ^= (int64_t) key[offset + length - rem + 5] << 40;
+        h64 ^= (uint64_t) key[offset + length - rem + 5] << 40;
     case 5:
-        h64 ^= (int64_t) key[offset + length - rem + 4] << 32;
+        h64 ^= (uint64_t) key[offset + length - rem + 4] << 32;
     case 4:
-        h64 ^= (int64_t) key[offset + length - rem + 3] << 24;
+        h64 ^= (uint64_t) key[offset + length - rem + 3] << 24;
     case 3:
-        h64 ^= (int64_t) key[offset + length - rem + 2] << 16;
+        h64 ^= (uint64_t) key[offset + length - rem + 2] << 16;
     case 2:
-        h64 ^= (int64_t) key[offset + length - rem + 1] << 8;
+        h64 ^= (uint64_t) key[offset + length - rem + 1] << 8;
     case 1:
-        h64 ^= (int64_t) key[offset + length - rem];
+        h64 ^= (uint64_t) key[offset + length - rem];
         h64 *= m64;
     }
 
@@ -137,22 +137,22 @@ int64_t hash2_64(const bytes &key, int32_t offset, int32_t length, int64_t seed)
     return h64;
 }
 
-static int64_t getblock(const bytes &key, int32_t offset, int32_t index)
+static uint64_t getblock(const bytes &key, uint32_t offset, uint32_t index)
 {
-    int32_t i_8 = index << 3;
-    int32_t blockOffset = offset + i_8;
-    return ((int64_t) key[blockOffset + 0] & 0xff) + (((int64_t) key[blockOffset + 1] & 0xff) << 8) +
-            (((int64_t) key[blockOffset + 2] & 0xff) << 16) + (((int64_t) key[blockOffset + 3] & 0xff) << 24) +
-            (((int64_t) key[blockOffset + 4] & 0xff) << 32) + (((int64_t) key[blockOffset + 5] & 0xff) << 40) +
-            (((int64_t) key[blockOffset + 6] & 0xff) << 48) + (((int64_t) key[blockOffset + 7] & 0xff) << 56);
+    uint32_t i_8 = index << 3;
+    uint32_t blockOffset = offset + i_8;
+    return ((uint64_t) key[blockOffset + 0] & 0xff) + (((uint64_t) key[blockOffset + 1] & 0xff) << 8) +
+            (((uint64_t) key[blockOffset + 2] & 0xff) << 16) + (((uint64_t) key[blockOffset + 3] & 0xff) << 24) +
+            (((uint64_t) key[blockOffset + 4] & 0xff) << 32) + (((uint64_t) key[blockOffset + 5] & 0xff) << 40) +
+            (((uint64_t) key[blockOffset + 6] & 0xff) << 48) + (((uint64_t) key[blockOffset + 7] & 0xff) << 56);
 }
 
-static int64_t rotl64(int64_t v, int32_t n)
+static uint64_t rotl64(uint64_t v, uint32_t n)
 {
     return ((v << n) | ((uint64_t)v >> (64 - n)));
 }
 
-static int64_t fmix(int64_t k)
+static uint64_t fmix(uint64_t k)
 {
     k ^= (uint64_t)k >> 33;
     k *= 0xff51afd7ed558ccdL;
@@ -163,23 +163,23 @@ static int64_t fmix(int64_t k)
     return k;
 }
 
-void hash3_x64_128(const bytes &key, int32_t offset, int32_t length, int64_t seed, std::array<int64_t,2> &result)
+void hash3_x64_128(const bytes &key, uint32_t offset, uint32_t length, uint64_t seed, std::array<uint64_t,2> &result)
 {
-    const int32_t nblocks = length >> 4; // Process as 128-bit blocks.
+    const uint32_t nblocks = length >> 4; // Process as 128-bit blocks.
 
-    int64_t h1 = seed;
-    int64_t h2 = seed;
+    uint64_t h1 = seed;
+    uint64_t h2 = seed;
 
-    int64_t c1 = 0x87c37b91114253d5L;
-    int64_t c2 = 0x4cf5ad432745937fL;
+    uint64_t c1 = 0x87c37b91114253d5L;
+    uint64_t c2 = 0x4cf5ad432745937fL;
 
     //----------
     // body
 
-    for(int32_t i = 0; i < nblocks; i++)
+    for(uint32_t i = 0; i < nblocks; i++)
     {
-        int64_t k1 = getblock(key, offset, i*2+0);
-        int64_t k2 = getblock(key, offset, i*2+1);
+        uint64_t k1 = getblock(key, offset, i*2+0);
+        uint64_t k2 = getblock(key, offset, i*2+1);
 
         k1 *= c1; k1 = rotl64(k1,31); k1 *= c2; h1 ^= k1;
 
@@ -196,27 +196,27 @@ void hash3_x64_128(const bytes &key, int32_t offset, int32_t length, int64_t see
     // Advance offset to the unprocessed tail of the data.
     offset += nblocks * 16;
 
-    int64_t k1 = 0;
-    int64_t k2 = 0;
+    uint64_t k1 = 0;
+    uint64_t k2 = 0;
 
     switch(length & 15)
     {
-    case 15: k2 ^= ((int64_t) key[offset+14]) << 48;
-    case 14: k2 ^= ((int64_t) key[offset+13]) << 40;
-    case 13: k2 ^= ((int64_t) key[offset+12]) << 32;
-    case 12: k2 ^= ((int64_t) key[offset+11]) << 24;
-    case 11: k2 ^= ((int64_t) key[offset+10]) << 16;
-    case 10: k2 ^= ((int64_t) key[offset+9]) << 8;
-    case  9: k2 ^= ((int64_t) key[offset+8]) << 0;
+    case 15: k2 ^= ((uint64_t) key[offset+14]) << 48;
+    case 14: k2 ^= ((uint64_t) key[offset+13]) << 40;
+    case 13: k2 ^= ((uint64_t) key[offset+12]) << 32;
+    case 12: k2 ^= ((uint64_t) key[offset+11]) << 24;
+    case 11: k2 ^= ((uint64_t) key[offset+10]) << 16;
+    case 10: k2 ^= ((uint64_t) key[offset+9]) << 8;
+    case  9: k2 ^= ((uint64_t) key[offset+8]) << 0;
         k2 *= c2; k2  = rotl64(k2,33); k2 *= c1; h2 ^= k2;
-    case  8: k1 ^= ((int64_t) key[offset+7]) << 56;
-    case  7: k1 ^= ((int64_t) key[offset+6]) << 48;
-    case  6: k1 ^= ((int64_t) key[offset+5]) << 40;
-    case  5: k1 ^= ((int64_t) key[offset+4]) << 32;
-    case  4: k1 ^= ((int64_t) key[offset+3]) << 24;
-    case  3: k1 ^= ((int64_t) key[offset+2]) << 16;
-    case  2: k1 ^= ((int64_t) key[offset+1]) << 8;
-    case  1: k1 ^= ((int64_t) key[offset]);
+    case  8: k1 ^= ((uint64_t) key[offset+7]) << 56;
+    case  7: k1 ^= ((uint64_t) key[offset+6]) << 48;
+    case  6: k1 ^= ((uint64_t) key[offset+5]) << 40;
+    case  5: k1 ^= ((uint64_t) key[offset+4]) << 32;
+    case  4: k1 ^= ((uint64_t) key[offset+3]) << 24;
+    case  3: k1 ^= ((uint64_t) key[offset+2]) << 16;
+    case  2: k1 ^= ((uint64_t) key[offset+1]) << 8;
+    case  1: k1 ^= ((uint64_t) key[offset]);
         k1 *= c1; k1  = rotl64(k1,31); k1 *= c2; h1 ^= k1;
     };
